backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
cohere==5.20.0
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.124.4
fastapi-cli==0.0.16
fastapi-cloud-cli==0.6.0
fastar==0.8.0
fastavro==1.12.1
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.12.0
greenlet==3.3.0
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.0
huggingface_hub==1.2.3
hyperframe==6.1.0
idna==3.11
Jinja2==3.1.6
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.0.354
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.2.1
langchain-text-splitters==1.0.0
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.0
langsmith==0.4.59
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
mdurl==0.1.2
multidict==6.7.0
mypy_extensions==1.1.0
numpy==1.26.4
openai==2.11.0
orjson==3.11.5
ormsgpack==1.12.0
packaging==23.2
portalocker==3.2.0
propcache==0.4.1
protobuf==6.33.2
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
python-dotenv==1.2.1
python-multipart==0.0.20
PyYAML==6.0.3
qdrant-client==1.16.2
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rich-toolkit==0.17.0
rignore==0.7.6
sentry-sdk==2.47.0
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.45
starlette==0.50.0
tenacity==8.5.0
tokenizers==0.22.1
tqdm==4.67.1
typer==0.20.0
typer-slim==0.20.0
types-requests==2.32.4.20250913
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uuid_utils==0.12.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
